+++ resym_core/src/pdb_types/union.rs
@@ -204,7 +204,7 @@ impl<'p> Union<'p> {
                     type_left,
                     type_right,
                     name: data.name,
-                    offset: data.offset,
+                    offset: data.offset as u32,
                     size: type_size,
                     access,
                 });
